+++ changelog.md
@@ -1,4 +1,9 @@
+## Changelog next version
+
+* Renamed method `shutdownSubscribers` in `DurableSubscriptonModel` to `shutdown`.  
+
 ## Changelog 0.5.1 (2021-01-07)
+
 * Renamed `org.occurrent.subscription.redis.spring.blocking.SpringSubscriptionPositionStorageForRedis` to `SpringRedisSubscriptionPositionStorage`.
 * Renamed `org.occurrent.subscription.mongodb.spring.reactor.ReactorMongoSubscription` to `ReactorMongoSubscriptionModel`.